Add password printing feature and fixed a minor keyboard navigation bug  (#44)
- Added a print password feature to the program. A user can now print
out a password on a printer. The printed password can include an
optional description and username.
- Improved data validation when reading from the settings file.
- Fixed a bug. The Main Window’s tab order was wrong because the tab
index was set on the Help button. The tab index does not need to be set
on this window because WPF correctly calculates the tab order.
- The text copied to the clip board is Unicode text instead of ASCII
text when the Copy Password to Clipboard button is pressed.

Co-authored-by: Benjamin Ellett <git@benellett.net>